# app/Http/Controllers/Api/AuthController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Auth;
use Illuminate\Support\Facades\Validator;
use Exception;
use Illuminate\Support\Facades\Cache;

/**
 * @OA\Info(
 *    title="Subscription Application",
 *    version="1.0.0",
 * )
 */
class AuthController extends Controller
{

    /**
         * @OA\Post(
         * path="/api/login",
         * summary="Sign in",
         * description="Login by email, password",
         * operationId="authLogin",
         * tags={"auth"},
         * @OA\RequestBody(
         *    required=true,
         *    description="Pass user credentials",
         *    @OA\JsonContent(
         *       required={"email","password"},
         *       @OA\Property(property="email", type="string", format="email", example="user1@mail.com"),
         *       @OA\Property(property="password", type="string", format="password", example="PassWord12345"),
         *       @OA\Property(property="persistent", type="boolean", example="true"),
         *    ),
         * ),
         * @OA\Response(
         *    response=422,
         *    description="Wrong credentials response",
         *    @OA\JsonContent(
         *       @OA\Property(property="message", type="string", example="Sorry, wrong email address or password. Please try again")
         *        )
         *     )
         * )
    */
    public function login(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'email'    => 'required|email',
            'password' => 'required'
        ]);

        if ($validator->fails()) return sendError('Validation Error.', $validator->errors(), 422);

        $credentials = $request->only('email', 'password');


        if (Auth::attempt($credentials)) {
            $user             = Auth::user();
            $success['name']  = $user->name;
            $success['token'] = $user->createToken('accessToken')->accessToken;

            Cache::store('file')->put($user->email, $user, 600); // 10 Minutes Caching
            Cache::store('file')->put($user->email,$success['token'], 600); // 10 Minutes


            return sendResponse($success, 'You are successfully logged in.');
        } else {
            return sendError('Unauthorised', ['error' => 'Unauthorised'], 401);
        }
    }


     /**
        * @OA\Post(
        * path="/api/register",
        * operationId="Register",
        * tags={"Register"},
        * summary="User Register",
        * description="User Register here",
        *     @OA\RequestBody(
        *         @OA\JsonContent(),
        *         @OA\MediaType(
        *            mediaType="multipart/form-data",
        *            @OA\Schema(
        *               type="object",
        *               required={"name","email", "password", "password_confirmation"},
        *               @OA\Property(property="name", type="text"),
        *               @OA\Property(property="email", type="text"),
        *               @OA\Property(property="password", type="password"),
        *               @OA\Property(property="password_confirmation", type="password")
        *            ),
        *        ),
        *    ),
        *      @OA\Response(
        *          response=201,
        *          description="Register Successfully",
        *          @OA\JsonContent()
        *       ),
        *      @OA\Response(
        *          response=200,
        *          description="Register Successfully",
        *          @OA\JsonContent()
        *       ),
        *      @OA\Response(
        *          response=422,
        *          description="Unprocessable Entity",
        *          @OA\JsonContent()
        *       ),
        *      @OA\Response(response=400, description="Bad request"),
        *      @OA\Response(response=404, description="Resource Not Found"),
        * )
    */
    public function register(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'name'     => 'required',
            'email'    => 'required|email|unique:users',
            'password' => 'required|min:8'
        ]);

        if ($validator->fails()) return sendError('Validation Error.', $validator->errors(), 422);

        try {
            $user = User::create([
                'name'     => $request->name,
                'email'    => $request->email,
                'password' => bcrypt($request->password)
            ]);

            $success['name']  = $user->name;
            $message          = 'Yay! A user has been successfully created.';
            $success['token'] = $user->createToken('accessToken')->accessToken;
        } catch (Exception $e) {
            $success['token'] = [];
            $message          = 'Oops! Unable to create a new user.';
        }

        return sendResponse($success, $message);
    }



    public function test(Request $request){
        if(Auth::user()->hasRole('test-role')){
            return response()->json(['message' => 'You have called the test authenticated api'], 200);
        }else{
            return response()->json(['message' => 'Unauthorized'], 403);
        }
    }




     /**
         * @OA\Get(
         * path="/api/v1/access-cached-token",
         * summary="access-cached-token",
         * description="access-cached-token",
         * operationId="access-cached-token",
         * tags={"access-cached-token"},
         * @OA\RequestBody(
         *    required=true,
         *    description="Pass user credentials",
         *    @OA\JsonContent(
         *
         *    ),
         * ),
         * @OA\Response(
         *    response=422,
         *    description="Wrong  response",
         *    @OA\JsonContent(
         *       @OA\Property(property="message", type="string", example="Sorry, Please try again")
         *        )
         *     )
         * )
    */
    public function getCachedToken(Request $request)
    {
        $value = Cache::get($request->user->email);

        if($value)
            return response()->json(["cachedAccessToken", Cache::get($request->user->email)]);

    }
}